backend/requirements.txt
annotated-doc==0.0.4
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cuda-bindings==13.2.0
cuda-pathfinder==1.5.1
cuda-toolkit==13.0.2
curl_cffi==0.13.0
dnspython==2.8.0
exceptiongroup==1.3.1
feedparser==6.0.12
filelock==3.25.2
frozendict==2.4.7
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.9.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multitasking==0.0.12
networkx==3.4.2
numpy==2.2.6
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
packaging==26.0
pandas==2.3.3
peewee==4.0.4
pillow==12.2.0
platformdirs==4.9.4
protobuf==7.34.1
pycparser==3.0
Pygments==2.20.0
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==14.3.3
safetensors==0.7.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
sympy==1.14.0
tokenizers==0.22.2
torch==2.11.0
torchaudio==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==5.5.0
triton==3.6.0
typer==0.24.1
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
websockets==16.0
yfinance==1.2.0
